+++ components/src/Box/Box.story.js
@@ -20,8 +20,8 @@ storiesOf("Box", module)
     </Box>
   ))
   .add("With a responsive width", () => (
-    <Box p="x3" width={{ small: 1, medium: 1 / 2, large: 1 / 4 }} bg="whiteGrey">
-      Full width on small screens, 1/2 width on medium and 1/4 width on large
+    <Box p="x3" width={{ extraSmall: 1, small: 1 / 2, medium: 1 / 4 }} bg="whiteGrey">
+      Full width on extra small screens, 1/2 width on small and 1/4 width on medium
     </Box>
   ))
   .add("With padding", () => (